rearrange DataTransfer directory and split navigation-related classes into new Navigation feature
