+++ WoWsShipBuilder.Common/Features/AccelerationCharts/CustomAccelerationDataDialog.razor
@@ -2,7 +2,7 @@
 @using WoWsShipBuilder.DataStructures.Ship
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
 @using System.Collections.Immutable
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 @inject ILocalizer Localizer
 
 <MudDialog>
