+++ WoWsShipBuilder.Common/Features/AccelerationCharts/ShipBuildContainerSelectionDialog.razor
@@ -1,7 +1,7 @@
 ﻿@using System.Collections.Specialized
 @using System.Collections.ObjectModel
 @using DynamicData
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
 
 @implements IDisposable
