+++ WoWsShipBuilder.Common/Features/BallisticCharts/ShipAndShellSelectionDialog.razor
@@ -10,7 +10,7 @@
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
 @using WoWsShipBuilder.Infrastructure.Utility
 @using System.Collections.Immutable
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 @implements IDisposable
 @inject ILocalizer Localizer
