+++ WoWsShipBuilder.Common/Features/Builds/Components/BuildConfigurationDialog.razor
@@ -4,8 +4,8 @@
 @using Microsoft.Extensions.Options
 @using WoWsShipBuilder.DataStructures.Ship
 @using WoWsShipBuilder.Features.Builds
-@using WoWsShipBuilder.Features.DataTransfer
 @using WoWsShipBuilder.Features.LinkShortening
+@using WoWsShipBuilder.Features.Navigation
 @using WoWsShipBuilder.Features.ShipStats.Components
 @using WoWsShipBuilder.Infrastructure.ApplicationData
 @using WoWsShipBuilder.Infrastructure.GameData
