+++ WoWsShipBuilder.Common/Features/Builds/Components/BuildConfigurationDialogHelper.razor
@@ -1,5 +1,5 @@
 ﻿@using Microsoft.Extensions.Options
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 @using WoWsShipBuilder.Infrastructure.ApplicationData
 
 @inject IDialogService DialogService
