+++ WoWsShipBuilder.Common/Features/Builds/ShipBuildViewModel.cs
@@ -2,7 +2,7 @@
 using WoWsShipBuilder.DataStructures.Modifiers;
 using WoWsShipBuilder.DataStructures.Ship;
 using WoWsShipBuilder.Features.DataContainers;
-using WoWsShipBuilder.Features.DataTransfer;
+using WoWsShipBuilder.Features.Navigation;
 using WoWsShipBuilder.Features.ShipStats.ViewModels;
 using WoWsShipBuilder.Infrastructure.ApplicationData;
 using WoWsShipBuilder.Infrastructure.Utility;
