+++ WoWsShipBuilder.Common/Features/FiringAngleDiagram/FiringAngleDialog.razor
@@ -1,5 +1,5 @@
 ﻿@using WoWsShipBuilder.Infrastructure.Localization.Resources
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Infrastructure.DataTransfer
 @implements IAsyncDisposable
 
 @inject IJSRuntime Runtime
