+++ WoWsShipBuilder.Common/Features/ShipComparison/ShipComparison.razor
@@ -36,7 +36,7 @@
 @using static WoWsShipBuilder.Features.ShipComparison.GridSections.BombersFragment
 @using static WoWsShipBuilder.Features.ShipComparison.GridSections.BombsFragment
 @using System.Collections.Immutable
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 
 @inherits ReactiveComponentBase<ShipComparisonViewModel>
