+++ WoWsShipBuilder.Common/Features/ShipSelection/FittingToolSelector.razor
@@ -7,7 +7,7 @@
 @using WoWsShipBuilder.DataStructures.Ship
 @using WoWsShipBuilder.Features.Settings
 @using System.Collections.Immutable
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 
 @inject ILocalizer Localizer
 @inject IHostEnvironment Environment
