+++ WoWsShipBuilder.Common/Features/ShipSelection/ShipSelectionDialog.razor
@@ -1,6 +1,6 @@
 ﻿@using System.Collections.Specialized
 @using System.Collections.ObjectModel
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
 
 @inject ILocalizer Localizer
