+++ WoWsShipBuilder.Common/Features/ShipSelection/ShipSelector.razor
@@ -16,7 +16,7 @@
 @using WoWsShipBuilder.Infrastructure.Utility
 @using System.Collections.Immutable
 @using System.Text.Json
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 
 @implements IAsyncDisposable
 @inject ILocalizer Localizer
