+++ WoWsShipBuilder.Common/Features/ShipStats/Components/ShipStatsPanel.razor
@@ -5,10 +5,10 @@
 @using WoWsShipBuilder.DataStructures.Ship.Components
 @using System.ComponentModel.DataAnnotations
 @using WoWsShipBuilder.Features.DataContainers
-@using WoWsShipBuilder.Features.DataTransfer
 @using WoWsShipBuilder.Features.FiringAngleDiagram
 @using WoWsShipBuilder.Features.Settings
 @using WoWsShipBuilder.Infrastructure
+@using WoWsShipBuilder.Infrastructure.DataTransfer
 @using WoWsShipBuilder.Infrastructure.GameData
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
 @using WoWsShipBuilder.Infrastructure.Metrics
