+++ WoWsShipBuilder.Common/Features/ShipStats/Components/ShipStatsTab.razor
@@ -18,7 +18,8 @@
 @using System.Diagnostics
 @using System.Collections.ObjectModel
 @using WoWsShipBuilder.DataStructures.Modifiers
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
+@using WoWsShipBuilder.Infrastructure.DataTransfer
 
 @inject IHostEnvironment Environment
 @inject ILocalizer Localizer
