+++ WoWsShipBuilder.Common/Features/ShipStats/ShipStats.razor
@@ -10,7 +10,7 @@
 @using WoWsShipBuilder.Infrastructure.Metrics
 @using WoWsShipBuilder.Infrastructure.Utility
 @using System.Collections.Immutable
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 
 @implements IDisposable
