+++ WoWsShipBuilder.Common/Features/StartPage/Index.razor
@@ -1,7 +1,7 @@
 @page "/"
 @using System.Collections.ObjectModel
 @using System.Collections.Specialized
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
 @using WoWsShipBuilder.Infrastructure.Utility
 @inject RefreshNotifierService NotifierService
