+++ WoWsShipBuilder.Common/Infrastructure/DataTransfer/GunDataContainer.cs
@@ -1,3 +1,3 @@
-﻿namespace WoWsShipBuilder.Features.DataTransfer;
+﻿namespace WoWsShipBuilder.Infrastructure.DataTransfer;
 
 public sealed record GunDataContainer(decimal HPos, decimal VPos, decimal BaseAngle, decimal[] Sector, decimal[][] DeadZones);
