+++ WoWsShipBuilder.Common/Infrastructure/Localization/ILocalizationProvider.cs
@@ -1,4 +1,4 @@
-﻿using WoWsShipBuilder.Features.DataTransfer;
+﻿using WoWsShipBuilder.Infrastructure.DataTransfer;
 using WoWsShipBuilder.Infrastructure.GameData;
 
 namespace WoWsShipBuilder.Infrastructure.Localization;
