+++ WoWsShipBuilder.Common/Infrastructure/Utility/ServiceCollectionExtensions.cs
@@ -3,8 +3,8 @@ using Microsoft.Extensions.DependencyInjection;
 using MudBlazor;
 using MudBlazor.Services;
 using WoWsShipBuilder.Features.BallisticCharts;
-using WoWsShipBuilder.Features.DataTransfer;
 using WoWsShipBuilder.Features.LinkShortening;
+using WoWsShipBuilder.Features.Navigation;
 using WoWsShipBuilder.Features.Settings;
 using WoWsShipBuilder.Features.ShipStats;
 using WoWsShipBuilder.Infrastructure.Localization;
