+++ WoWsShipBuilder.Desktop/Features/Settings/SettingsComponent.razor
@@ -20,7 +20,7 @@
 @using System.IO
 @using System.Text.Json
 @using Avalonia.Platform.Storage
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Infrastructure.DataTransfer
 
 @implements IDisposable
 @inject ILocalizer Localizer
