+++ WoWsShipBuilder.Web/Features/AdminPage/AdminPage.razor
@@ -5,7 +5,7 @@
 @using WoWsShipBuilder.Infrastructure.Localization
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
 @using System.Security.Claims
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Infrastructure.DataTransfer
 @using WoWsShipBuilder.Web.Features.BetaAccess
 
 @inject ILocalizer Localizer
