+++ WoWsShipBuilder.Web/Features/TestPage/TestPage.razor
@@ -1,10 +1,10 @@
 ﻿@page "/test"
 @using WoWsShipBuilder.DataStructures
 @using WoWsShipBuilder.DataStructures.Ship
-@using WoWsShipBuilder.Features.DataTransfer
 @using WoWsShipBuilder.Features.FiringAngleDiagram
 @using WoWsShipBuilder.Features.Settings
 @using WoWsShipBuilder.Infrastructure.ApplicationData
+@using WoWsShipBuilder.Infrastructure.DataTransfer
 @using WoWsShipBuilder.Infrastructure.GameData
 @using WoWsShipBuilder.Infrastructure.Localization
 @using WoWsShipBuilder.Infrastructure.Localization.Resources
