+++ WoWsShipBuilder.Web/Features/TestPage/TestPage2.razor
@@ -1,7 +1,7 @@
 ﻿@page "/test2"
 @using WoWsShipBuilder.DataStructures
 @using WoWsShipBuilder.Features.Builds.Components
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Features.Navigation
 @using WoWsShipBuilder.Infrastructure.ApplicationData
 
 <MudRadioGroup @bind-SelectedOption="selectedTier">
