+++ WoWsShipBuilder.Web/Features/UserSettings/UserSettings.razor
@@ -9,7 +9,7 @@
 @using WoWsShipBuilder.Infrastructure.Utility
 @using WoWsShipBuilder.Web.Features.BetaAccess
 @using System.Text.Json
-@using WoWsShipBuilder.Features.DataTransfer
+@using WoWsShipBuilder.Infrastructure.DataTransfer
 
 @implements IDisposable
 @implements IAsyncDisposable
