+++ WoWsShipBuilder.Web/Infrastructure/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
-﻿using WoWsShipBuilder.Features.DataTransfer;
-using WoWsShipBuilder.Infrastructure.ApplicationData;
+﻿using WoWsShipBuilder.Infrastructure.ApplicationData;
 using WoWsShipBuilder.Infrastructure.HttpClients;
 using WoWsShipBuilder.Infrastructure.Utility;
 using WoWsShipBuilder.Web.Features.Authentication;
@@ -7,6 +6,8 @@ using WoWsShipBuilder.Web.Features.BetaAccess;
 using WoWsShipBuilder.Web.Infrastructure.Data;
 
 namespace WoWsShipBuilder.Web.Infrastructure;
+
+using WoWsShipBuilder.Infrastructure.DataTransfer;
 
 public static class ServiceCollectionExtensions
 {
